refactor: extract shared dialog lifecycle and consolidate error parsing
Phase 2 of validated technical-debt cleanup:

- Extract createDialogLifecycle helper (dialog-lifecycle.js) shared by
  import-panel and sharing-centre; import-panel now also closes on Escape
  (a11y fix), matching sharing-centre
- Export parseErrorResponse as a standalone from api-core and reuse it in
  background.js, removing the duplicated parseApiError. Aligns field-order
  to the consolidated message || error contract
- Remove the coupled total/hasMore compat shim: drop legacy top-level
  pagination fields from GetPagesResponseSchema and the matching fallbacks
  in normalizePagination. The backend now emits only the pagination block;
  updated the normalizer test accordingly
