+++ src/validators.js
@@ -80,10 +80,6 @@ export const PageSchema = z.strictObject({
  */
 export const GetPagesResponseSchema = z.object({
   pages: z.array(PageSchema),
-  // Legacy fields (deprecated)
-  total: z.number().int().nonnegative().optional(),
-  hasMore: z.boolean().optional(),
-  // New pagination fields
   pagination: z.object({
     nextCursor: z.string().nullable(),
     hasNextPage: z.boolean()
